Support batch transactions with single tan, fix overflow
- Remove tan from batchfile, can now be entered during fileupload
- Correct datatype when parsing the transaction amount
- Improved error messages

+++ project/accounts/new_transaction_multiple.php
@@ -17,14 +17,16 @@ if (empty($_SESSION["account_id"]))
 	die("Please choose an account");
 
 $account_id = $_SESSION["account_id"];
-if (isset($_FILES['transactionsCSV'])) {
-	$file = $_FILES['transactionsCSV'];
+if (isset($_FILES["transactionsCSV"]) && isset($_POST["tan"])) {
+	$file = $_FILES["transactionsCSV"];
+	$tan = $_POST["tan"];
+	$tan = preg_replace("([^a-zA-Z0-9+\/])", '', $tan);
 	$name = session_id();
 	$target_file = getPageAbsolute("uploads") . $name;
 	$ctransact = getPageAbsolute("ctransact");
 	if (move_uploaded_file($file['tmp_name'], $target_file)) {
 		echo "<div class='success'>Fileupload successful! Starting batch processing...<br />";
-		$cmdln = "$ctransact '$account_id' '$target_file'";
+		$cmdln = "$ctransact '$account_id' '$tan' '$target_file'";
 		exec($cmdln, $cmdout);
 		unlink($target_file);
 		foreach ($cmdout as $line) {
@@ -47,13 +49,31 @@ if (isset($_FILES['transactionsCSV'])) {
 		<li>No quoting of whitespace required</li>
 		<li>The complete input between two commas gets treated as value</li>
 		<li>Example:</li>
-		<p class="simple-text-big">DST_ACCOUNT,AMOUNT,DESCRIPTION,TAN</p>
+		<p class="simple-text-big">DST_ACCOUNT,AMOUNT,DESCRIPTION</p>
 	</ul>
 </p>
 <br />
 <p class="simple-text">Note: All Transactions over 10,000 will require manual approval by an employee</p>
 <form id="uploadForm" method="post" enctype="multipart/form-data" class="simple-text">
-	Select transaction file to upload:
-	<input type="file" name="transactionsCSV" id="transactionsCSV">
-	<button type="button" onclick="uploadFile()" class="simpleButton">Upload</button>
+	<div class="transaction-container">
+		<div class="formRow">
+			<div class="formLeftColumn">
+				<label for="transactionsCSV" class="simple-label">Transaction file </label>
+			</div>
+			<div class="formRightColumn">
+				<input type="file" name="transactionsCSV" id="transactionsCSV"><br>
+			</div>
+		</div>
+		<div class="formRow">
+			<div class="formLeftColumn">
+				<label for="tan" class="simple-label">TAN </label>
+			</div>
+			<div class="formRightColumn">
+				<input type="text" name="tan" id="tan" placeholder="TAN"><br>
+			</div>
+		</div>
+		<div class="button-container">
+			<button type="button" onclick="uploadFile()" class="simpleButton">Upload</button>
+		</div>
+	</div>
 </form>
